Project files/requirements.txt
streamlit
fastapi
uvicorn
requests
python-dotenv
sentence-transformers
pydantic-settings
pinecone-client==3.0.0
scikit-learn
pandas
matplotlib